Improve keyspace name verification
